libedgedns/conductor: improve pending query cancellation
When starting a new outbound query with Conductor, the timetable
returns a handle pointing to a queue of waiting queries.
This handle effectively owns the pending query, and clears it
when it goes out of scope.

This fixes a problem when caller cancelled a future that included
a waiting conductor query, since the future wouldn't be able to finish,
the query would be stuck in the pending queue forever.

The exchangers also support fallbacks now, in case of UDP it falls
back on TCP because of truncation. The TCP exchanger falls back to UDP
if all connection attempts fail.