libedgedns: moved connection reuse to right place, concurrent tries in conductor
The retries are now staggered based on the expected RTT in conductor, instead of
being issued sequentially. The conductor doesn't reshuffle origin selection for
connection reuse, this is now done in the origin (if it prefers so). The conductor
processes origin selection now as-is, and reuses connection if any given try has
a pending connection available.

The origin now has access to Timetable, and connection metrics to make better decisions
on upstream selection.