libedgedns: refactor conductor to add fast retries and TLS
The commits changes the behavior with conductor retries. The algorithm now is
to get a sorted list of possible upstreams from Origin. It then plans the attempts
based on expected RTT of the upstream. So the first query is sent at T+0, and say
the expected RTT is 100ms. The next query is planned at T+150 (expected RTT + buffer),
or when the first query completes with an error. This improves a situation when origins
refuse TCP/TLS immediately, so that the next retry doesn't have to wait for the full grace period.
This is now the default Exchanger implementation, instead of preselected TCP or UDP exchanger.

The conductor also adds support for TLS for outbound connections to port 853.
The tracing now shows UDP fallbacks as separate spans instead of just annotations.