libedgedns: unified FramedStream for all transport protocols, enable DoT on server side
The FramedStream now abstracts any stream of DNS messages, which allows for a common processing
functions in Server and Conductor. The Server now enables DoT processing, which required a change
in the configuration. In addition to `listen = "addr:port"`, the configuration also accepts enumerated
tables:

```
[network.listen.addr1]
address = "0.0.0.0:53"

[network.listen.addr2]
address = "0.0.0.0:853"
tls = ["cert.p12", "password"]
```

The listeners are symbolically named so that their configuration can be linked to named fds from supervisor later.