sandbox: integrate into core
Based on WebAssembly, the sandbox provides some kind of isolation
between host and guest, also between each guest apps. Current wasmer
is used as a runtime, cranelift is the underlying compiler.

Only "filesystem" backend is supported for loading guest apps at the
moment. All the apps needs to be put together in one directory and
will be reloaded whenever the mtime changed.