libedgedns: fix future registration in sandbox::run()
The `sandbox::run(instance)` registers guest futures immediately in the call,
so they are tied to the task that calls the function. But the guest futures
could be polled from a different task (if the result of `sandbox::run()` is spawned
as a new task. In that case, the spawned task would never be notified.
Ideally the guest futures would be registered lazily to avoid the problem of
binding them to a wrong task altogether.