sandbox: poll host futures directly, add BufferedStream and example
The host futures are now not spawned as separate tasks, but polled
directly by guest. This is more convenient as we don't have to track which
task spawned the task, and also cancellation is implicit when the guest
stops polling the future.

Added BufferedStream implementation that roughly follows BufRead in guest.
Added an example for memcache protocol reader using the BufferedStream.