libedgedns/sandbox: support loading apps from a KV store
This allows loading apps dynamically based on the contents of a manifest.
A manifest is a value in the KV that lists all the enabled apps and their HMAC signatures.
The loader uses the manifest to figure out which apps need to be reloaded or removed.
When an app signature appears or changes, loader will load the app from the KV and
verify the signature. If it passes, it will load it and replace current instance.
When an app disappears from the manifest, loader will remove it from instance map.